+++ cnccode/data_classes/DBEPassword.inc.php
@@ -7,6 +7,19 @@ require_once($cfg["path_gc"] . "/DBEntity.inc.php");
 
 class DBEPassword extends DBEntity
 {
+
+    const passwordID = "passwordID";
+    const customerID = "customerID";
+    const username = "username";
+    const password = "password";
+    const notes = "notes";
+    const level = "level";
+    const URL = 'URL';
+    const archivedAt = 'archivedAt';
+    const archivedBy = 'archivedBy';
+    const serviceID = 'serviceID';
+    const encrypted = 'encrypted';
+
     /**
      * calls constructor()
      * @access public
@@ -18,14 +31,117 @@ class DBEPassword extends DBEntity
     {
         parent::__construct($owner);
         $this->setTableName("password");
-        $this->addColumn("passwordID", DA_ID, DA_NOT_NULL, "pas_passwordno");
-        $this->addColumn("customerID", DA_ID, DA_NOT_NULL, "pas_custno");
-        $this->addColumn("username", DA_STRING, DA_ALLOW_NULL, "pas_username");
-        $this->addColumn("service", DA_STRING, DA_ALLOW_NULL, "pas_service");
-        $this->addColumn("password", DA_STRING, DA_ALLOW_NULL, "pas_password");
-        $this->addColumn("notes", DA_STRING, DA_ALLOW_NULL, "pas_notes");
+        $this->addColumn(
+            self::passwordID,
+            DA_ID,
+            DA_NOT_NULL,
+            "pas_passwordno"
+        );
+        $this->addColumn(
+            self::customerID,
+            DA_ID,
+            DA_NOT_NULL,
+            "pas_custno"
+        );
+        $this->addColumn(
+            self::username,
+            DA_STRING,
+            DA_ALLOW_NULL,
+            "pas_username"
+        );
+        $this->addColumn(
+            self::password,
+            DA_STRING,
+            DA_ALLOW_NULL,
+            "pas_password"
+        );
+        $this->addColumn(
+            self::notes,
+            DA_STRING,
+            DA_ALLOW_NULL,
+            "pas_notes"
+        );
+
+        $this->addColumn(
+            self::level,
+            DA_INTEGER,
+            DA_NOT_NULL
+        );
+
+        $this->addColumn(
+            self::URL,
+            DA_STRING,
+            DA_ALLOW_NULL
+        );
+
+        $this->addColumn(
+            self::archivedAt,
+            DA_DATETIME,
+            DA_ALLOW_NULL
+        );
+
+        $this->addColumn(
+            self::archivedBy,
+            DA_STRING,
+            DA_ALLOW_NULL
+        );
+
+        $this->addColumn(
+            self::serviceID,
+            DA_INTEGER,
+            DA_ALLOW_NULL
+        );
+
+        $this->addColumn(
+            self::encrypted,
+            DA_INTEGER,
+            DA_ALLOW_NULL
+        );
+
         $this->setAddColumnsOff();
         $this->setPK(0);
+    }
+
+    public function getRowsByCustomerIDAndPasswordLevel($customerID,
+                                                        $passwordLevel,
+                                                        $archived = false
+    )
+    {
+        $this->setMethodName('getRowsByCustomerIDAndPasswordLevel');
+        if ($customerID == '') {
+            $this->raiseError('customerID not set');
+        }
+        if ($passwordLevel == '') {
+            $this->raiseError('passwordLevel not set');
+        }
+
+        $this->setQueryString(
+            "SELECT " . $this->getDBColumnNamesAsString() .
+            " FROM " . $this->getTableName() .
+            " WHERE " . $this->getDBColumnName(self::customerID) . "=" . $customerID .
+            " AND " . $this->getDBColumnName(self::level) . " <= " . $passwordLevel .
+            " and (" . $this->getDBColumnName(self::archivedBy) . ' is null or  ' . $this->getDBColumnName(
+                self::archivedBy
+            ) . ' = "" )'
+        );
+        return (parent::getRows());
+    }
+
+    public function getArchivedRowsByCustomerID($customerID)
+    {
+        $this->setMethodName('getArchivedRowsByCustomerID');
+        if ($customerID == '') {
+            $this->raiseError('customer ID not set');
+        }
+        $this->setQueryString(
+            "SELECT " . $this->getDBColumnNamesAsString() .
+            " FROM " . $this->getTableName() .
+            " WHERE " . $this->getDBColumnName(self::customerID) . " = " . $customerID .
+            " and " . $this->getDBColumnName(self::archivedBy) . ' is not null and  ' . $this->getDBColumnName(
+                self::archivedBy
+            ) . ' <> "" '
+        );
+        return (parent::getRows());
     }
 }