+++ htdocs/config.inc.php
@@ -88,6 +88,41 @@ define(
     'c:\\keys\\user-private.pub'
 );
 
+define(
+    'PASSWORD_ENCRYPTION_PRIVATE_KEY',
+    'c:\\keys\\passwordPrivate.pem'
+);
+
+define(
+    'PASSWORD_ENCRYPTION_PUBLIC_KEY',
+    'c:\\keys\\passwordPublic.pub'
+);
+
+define(
+    'PASSWORD_PASSPHRASE',
+    "adductor scincoid glue aviation sought beaker"
+);
+
+
+define(
+    'LABTECH_DB_HOST',
+    '192.168.33.64'
+);
+
+define(
+    'LABTECH_DB_NAME',
+    'labtech'
+);
+
+define(
+    'LABTECH_DB_USERNAME',
+    'cnccrmuser'
+);
+
+define(
+    'LABTECH_DB_PASSWORD',
+    'kj389fj29fjh'
+);
 
 $onPavilionWebServer = false;
 $GLOBALS['php7'] = true;
@@ -958,7 +993,10 @@ define(
     'mailshot8Flag'
 );    // GSC statement contact flag column
 
-define('CONFIG_HEADER_INVOICE_CONTACT', 'mailshot2Flag');     // Customer contact to send invoices to
+define(
+    'CONFIG_HEADER_INVOICE_CONTACT',
+    'mailshot2Flag'
+);     // Customer contact to send invoices to
 
 define(
     'CONFIG_HEADER_DAILY_OPEN_SR_REPORT',